3-print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -4,7 +4,7 @@
 
 /**
  * print_all - prints anything
- * @format:list of arguments passed to the function
+ * @format: list of arguments passed to the function
  */
 void print_all(const char * const format, ...)
 {
@@ -28,7 +28,7 @@ void print_all(const char * const format, ...)
 					printf("%s%d", sep, va_arg(list, int));
 					break;
 				case 'f':
-					printf("%s%f", sep, va_arg(list, int));
+					printf("%s%f", sep, va_arg(list, double));
 					break;
 				case 's':
 					str = va_arg(list, char *);
